+++ Grids/Hexagonal/Hexagon.cs
@@ -8,134 +8,42 @@ namespace Foundation.Grids.Hexagonal
 	/// Implementation following the guidelines at https://www.redblobgames.com/grids/hexagons/implementation.html
 	/// </summary>
 	[Serializable]
-	public class Hexagon : IEquatable<Hexagon>
+	public class Hexagon : MonoBehaviour
 	{
+		#region Inspector Fields
+		private float _size = 1f;
+		private HexagonCoordinates _coordinates = HexagonCoordinates.Zero;
+		#endregion
+
 		#region Properties
 		public float Size => _size;
+		public HexagonCoordinates Coordinates => _coordinates;
 
-		public int Q { get; }
-		public int R { get; }
-		public int S { get; }
-
-		public float Width
-			=> Mathf.Sqrt(3) * Size;
-
-		public float Height
-			=> 2 * Size;
-
-		public HexagonTile HexagonTile { get; set; }
+		public float Width => Mathf.Sqrt(3) * Size;
+		public float Height => 2 * Size;
 		#endregion
 
-		#region Fields
-		protected float _size = 1f;
-		#endregion
-
-		#region Constructors
-		public Hexagon(int q, int r, int s)
+		#region Life Cycle
+		public void Initialize(HexagonCoordinates coordinates)
 		{
-			if (q + r + s != 0)
-				throw new Exception("Hexagon coordinates don't add up to 0.");
-
-			Q = q;
-			R = r;
-			S = s;
+			_coordinates = coordinates;
 		}
 		#endregion
 
-		#region Equality Methods
-		public static bool operator ==(Hexagon a, Hexagon b)
-			=> a.Q == b.Q && a.R == b.R && a.S == b.S;
-
-		public static bool operator !=(Hexagon a, Hexagon b)
-			=>  !(a == b);
-
-		public bool Equals(Hexagon other)
-			=> this == other;
-
-		public override bool Equals(object o)
+		#region Methods
+		public void SetCoordinates(HexagonCoordinates coordinates)
 		{
-			if (o == null)
-				return false;
-
-			Hexagon hexagon = o as Hexagon;
-
-			return Equals(hexagon);
+			_coordinates = coordinates;
 		}
 
-		public override int GetHashCode()
+		public int GetDirectionFromTile(Hexagon hexagon)
 		{
-			int hashCode = -1013937445;
-			hashCode = hashCode * -1521134295 + Q.GetHashCode();
-			hashCode = hashCode * -1521134295 + R.GetHashCode();
-			hashCode = hashCode * -1521134295 + S.GetHashCode();
-			return hashCode;
+			return HexagonalDirections.Get(HexagonCoordinates.Subtract(Coordinates, hexagon.Coordinates).Normalized().ToVector3Int());
 		}
-		#endregion
-
-		#region Arithmetic Methos
-		public static Hexagon Add(Hexagon a, Hexagon b)
-			=> new Hexagon(a.Q + b.Q, a.R + b.R, a.S + b.S);
-
-		public static Hexagon Add(Hexagon a, Vector3Int vector)
-			=> Add(a, new Hexagon(vector.x, vector.y, vector.z));
-
-		public static Hexagon Subtract(Hexagon a, Hexagon b)
-			=>  new Hexagon(a.Q - b.Q, a.R - b.R, a.S - b.S);
-
-		public static Hexagon Subtract(Hexagon a, Vector3Int vector)
-			=> Subtract(a, new Hexagon(vector.x, vector.y, vector.z));
-
-		public static Hexagon Multiply(Hexagon a, int k)
-			=> new Hexagon(a.Q * k, a.R * k, a.S * k);
-		#endregion
-
-		#region Distance Methods
-		public static int Length(Hexagon hexagon)
-			=> Mathf.RoundToInt(Math.Abs(hexagon.Q) + Math.Abs(hexagon.R) + Math.Abs(hexagon.S) / 2);
-
-		public int Length()
-			=> Length(this);
-
-		public static int Distance(Hexagon a, Hexagon b)
-			=> Length(Subtract(a, b));
-
-		public Hexagon Normalized()
-			=> Normalized(this);
-
-		public static Hexagon Normalized(Hexagon hexagon)
-		{
-			int normalizedQ = Mathf.Clamp(hexagon.Q, -1, 1);
-			int normalizedR = Mathf.Clamp(hexagon.R, -1, 1);
-			int normalizedS = Mathf.Clamp(hexagon.S, -1, 1);
-
-			return new Hexagon(normalizedQ, normalizedR, normalizedS);
-		}
-		#endregion
-
-		#region Neighbor Methods
-		public Vector3Int Direction(int direction /* 0 to 5 */)
-			=> HexagonalDirections.Get(direction);
-
-		public static Hexagon Neighbour(Hexagon hexagon, int direction)
-			=>  Add(hexagon, hexagon.Direction(direction));
 		#endregion
 
 		#region Position Conversion
-		public Vector3 ToWorldPosition()
-		{
-			float x = Size * Mathf.Sqrt(3) * Q + Mathf.Sqrt(3) / 2 * R;
-			float y = Size * 3 / 2 * R;
-
-			return new Vector3(x, 0, y);
-		}
+		public Vector3 ToWorldPosition() => Coordinates.ToWorldPosition(Size);
 		#endregion
-
-		#region Convertors
-		public Vector3Int ToVector3Int()
-			=> new Vector3Int(Q, R, S);
-
-		public override string ToString()
-			=> $"({Q}, {R}, {S})";
-		#endregion
-	} 
+	}
 }